I guess we don't need a dependency on libGDX
The (silly) Null annotations are gone and replaced with the Jetbrains Nullable (and NotNull) annotations. MathUtils was replaced with a one-liner instead of its nextPowerOfTwo() method. Collections wasn't used, but now a version in this repo holds the tableSize() method and dummy constant. We also duplicate GdxRuntimeException for now; it will be renamed after more classes are brought over.