Fix a weird bug in remove() on iterators for lists.
The former bug probably affected a fair amount of things, all rarely-used. Items were removed by value, instead of index, which managed to slip by compile-time type checks because an int index can be used for an int value, a float value, a long value, or boxed to be an Object value.