Big license cleanup!
This adds the AUTHORS, CONTRIBUTORS, and THANKS files, the last one because I do want to thank people who have indirectly contributed. I also updated the copyright to 2021 for all files; almost everything was added this year, and if it wasn't, it was probably heavily updated this year. The Apache license header is now on every source file.