Bring Base over from SquidSquad.
It provides base-system conversion for numbers to and from Strings. It can also handle joining and splitting arrays of numbers, including 2D arrays. I think it should be useful for several parts of jdkgdxds-interop, which needs to convert to and from String often. It has more uses than just in the serialization part of a project, so I'm putting it here.