Having an Ordered and a Collection-based constructor was a problem for ObjectList.
We avoid some compile-time errors by only keeping one, in this case the constructor that takes a Collection. More (incorrect) IDE warnings/errors are avoided by staying on Java 16, because SplittableRandom.nextInt(int) was handled quite badly in Java 17 and requires language level 17 to not warn. This last part only affects a test.