Some more tests, used as a rough benchmark...
With Vector2 keys (that have small integer values in their x and y), ObjectObjectCuckooMap currently outperforms ObjectObjectMap by over a factor of 2. This is quickly reversed if the Fibonacci mixer is used, which makes ObjectObjectMap about 30x faster than ObjectObjectCuckooMap, but slows it down slightly if the hashes are already high-quality.