Identity hashes don't need more randomization.
Benchmarking IdentityObjectMap here shows that it can be a bit faster than ObjectObjectMap when the keys use the default hashCode(), but needs to avoid casting/adding hashMultiplier to get that extra speed. I don't think including hashMultiplier helped the randomness at all, it just changed the order slightly on resize(), but slowed down place() a little.