When using mask instead of shift...
It is possible to get much faster point hashing, but it cannot use the default IDEA-generated hashCode() because just a multiply by 31 and add doesn't produce large enough hashes for this table. Cantor is also sometimes drastically slower than Rosenberg-Strong, and I have no clue why (it might have to do with how the shells are generated).