A starting size of 16 instead of ArrayList's 10...
Appeared to cause some problems measuring memory use. We should probably act like the JDK types as much as possible, so the default size is now 10. This is probably good for memory usage in the common case where the default is used because the amount of items could be either low or high.