Change the default hashMultiplier.
The new one seems to yield better behavior on collision rates and maximum pileup than the old golden-ratio-based multiplier. I'm trying to remember exactly how I calculated this one, but it's 2 to the 64 divided by the third harmonic number (pretty sure on this). As to why this particular multiplier does well... ???