Adjust hash resize change a little (tested).
PileupTest suggests that the new hash resize strategy does a little better with realistic String hashCodes and somewhat less well with unrealistic hashCodes with artificially lowered bit counts. It also does just as well in the AllGolden tests for both Vector2 and dictionary words, compared to the last resize strategy.