Start making some enum-based sets/maps.
These are in a separate package of the tests for now. I feel like these are too narrow in usage to be useful in the main JAR, but small enough to be easily copied from the tests to a user project. They might get their own dependency if there is interest.