Rename and bring over the original... No dice.
The CuckooHashMap without any changes fails to insert anything with a completely-colliding hashCode(). This is similar to the issue with the old libGDX cuckoo-hashed maps and sets, though much worse here. It also clearly invalidates the claim of "amortized O(1) put()" if it takes infinite time to put() the third key.