Very small hashCodes collide more often. So?
A test could fail in ObjectObjectMap because it was checking that certain behavior would work, even though that behavior (mutating a key that is in a value-hash-based Map) should never be allowed or facilitated. The failure was dependent on whether there was a collision between some very small hashCode() results (1 and 17), and those tend to collide now because the multipliers are just smaller. By the time you have enough items for this to be a problem, they won't all be colliding.