Evaluate another GWT opt.
I am not sure how Integer.highestOneBit() is implemented on GWT, but if it is like the rest of that class, it isn't nearly as fast as it could be using code available to modern JS. I checked this and it seems reasonable for our use cases. One quirk is that index 0 would have resized to an invalid 0-length array before, and resizes to a 1-length array now (which can hold index 0).